+++ CHANGELOG.md
@@ -3,9 +3,16 @@
 
 ## [Unreleased]
 
+### Added
+- Documented the Golem Public API
+- UP, DOWN, LEFT, RIGHT options to actions.press_key (thanks to danielmaddern)
+
+###
+
+
 ## [0.1.22] - 2017-09-27
 ### Added
-- Added clear action to golem.actions
+- Added 'clear' action to golem.actions
 
 ### Fixed
 - Missing js files in manifest.in